requirements.txt
edge_tts==6.1.12
Flask==3.0.3
funasr==1.1.6
PyMySQL==1.1.1
python_bcrypt==0.3.2
Requests==2.32.3
